notification pt 1

+++ Felicette el Gatonauta/Assets/Scripts/Ads/AdManager.cs
@@ -14,7 +14,7 @@ public class AdManager : MonoBehaviour, IUnityAdsListener
     string adToShow = "Rewarded_Android";
     RewardType adRewardType;
 
-    public float staminaReward;
+    public int staminaReward;
 
     private void Start()
     {
